requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.2.0
async-timeout==5.0.1
attrs==24.3.0
Bottleneck==1.4.2
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
deepsnap==0.2.1
dgl==0.9.1.post1
dlp-mpi==0.0.4
docopt==0.6.2
einops==0.8.1
frozenlist==1.5.0
fsspec==2024.6.1
grpcio==1.71.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
HyperPyYAML==1.2.2
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.4
joblib==1.4.2
jsonpickle==3.2.2
lazy-dataset==0.0.15
lightning-utilities==0.11.9
loguru==0.7.3
Markdown==3.8
MarkupSafe==2.1.3
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
multidict==6.1.0
munch==4.0.0
networkx==3.2.1
numexpr==2.10.1
numpy==1.26.4
olefile==0.47
packaging==24.2
pandas==2.2.3
Pillow==8.2.0
pip==24.2
propcache==0.2.0
protobuf==6.30.2
psutil==6.1.1
py-cpuinfo==9.0.0
pycparser==2.22
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
sacred==0.8.5
scikit-learn==1.6.1
scipy==1.13.1
setuptools==75.1.0
sh==2.0.7
six==1.16.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
torch==1.12.1
torch_geometric==2.5.2
torch_scatter==2.0.9
torch_sparse==0.6.13
torchaudio==0.12.1
torcheval==0.0.7
torchmetrics==1.4.0.post0
torchvision==0.13.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.44.0
yacs==0.1.8
yarl==1.18.0
zipp==3.21.0
zstandard==0.19
